@@ -1,17 +1,23 @@
-namespace multiplayerSnake
+﻿using Raylib_cs;
+
+namespace HelloWorld;
+
+class Program
 {
-    internal static class Program
+    public static void Main()
     {
-        /// <summary>
-        ///  The main entry point for the application.
-        /// </summary>
-        [STAThread]
-        static void Main()
+        Raylib.InitWindow(800, 480, "Hello World");
+
+        while (!Raylib.WindowShouldClose())
         {
-            // To customize application configuration such as set high DPI settings or default font,
-            // see https://aka.ms/applicationconfiguration.
-            ApplicationConfiguration.Initialize();
-            Application.Run(new Form1());
+            Raylib.BeginDrawing();
+            Raylib.ClearBackground(Color.White);
+
+            Raylib.DrawText("Hello, world!", 12, 12, 20, Color.Black);
+
+            Raylib.EndDrawing();
         }
+
+        Raylib.CloseWindow();
     }
 }
